Move raw data paths from each individual script to common spec_shared.C.  Move common gHcParms and THaApparatus construction to the spec_shared.C file.  Minor correction to DBASE/HMS/standard.database to point to correct BCM calibration for sp18 boiling runs.
