--- SCRIPTS/SHMS/TEST_STANDS/replay_pngcer_test_stand.C
@@ -1,108 +1,0 @@
-void replay_pngcer_test_stand(Int_t RunNumber=0, Int_t MaxEvent=0) {
-
-  // Get RunNumber and MaxEvent if not provided.
-  if(RunNumber == 0) {
-    cout << "Enter a Run Number (-1 to exit): ";
-    cin >> RunNumber;
-    if( RunNumber<=0 ) return;
-  }
-  if(MaxEvent == 0) {
-    cout << "\nNumber of Events to analyze: ";
-    cin >> MaxEvent;
-    if(MaxEvent == 0) {
-      cerr << "...Invalid entry\n";
-      exit;
-    }
-  }
-
-  // Create file name patterns.
-  const char* RunFileNamePattern = "shms_all_%05d.dat";
-  vector<TString> pathList;
-  pathList.push_back(".");
-  pathList.push_back("./raw");
-  pathList.push_back("./raw/../raw.copiedtotape");
-  pathList.push_back("./cache");
-
-  const char* ROOTFileNamePattern = "ROOTfiles/pngcer_replay_%d.root";
-  // Add variables to global list.
-  gHcParms->Define("gen_run_number", "Run Number", RunNumber);
-  gHcParms->AddString("g_ctp_database_filename", "DBASE/SHMS/standard.database");
-  gHcParms->Load(gHcParms->GetString("g_ctp_database_filename"), RunNumber);
-  gHcParms->Load(gHcParms->GetString("g_ctp_parm_filename"));
-  gHcParms->Load(gHcParms->GetString("g_ctp_kinematics_filename"), RunNumber);
-  gHcParms->Load("PARAM/TRIG/tshms.param");
-  gHcParms->Load("PARAM/SHMS/GEN/p_fadc_debug.param");
-
-  // Load the Hall C style detector map
-  gHcDetectorMap = new THcDetectorMap();
-  gHcDetectorMap->Load("MAPS/SHMS/DETEC/STACK/shms_stack.map");
-  
-  // Set up the equipment to be analyzed.
-  THaApparatus* SHMS = new THcHallCSpectrometer("P", "SHMS");
-  gHaApps->Add(SHMS);
-  // Add hodoscope to SHMS apparatus
-  THcCherenkov* ngcer = new THcCherenkov("ngcer", "Noble Gas Cherenkov");
-  SHMS->AddDetector(ngcer);
-  // Add hodoscope to SHMS apparatus
-  THcHodoscope* hod = new THcHodoscope("hod", "Hodoscope");
-  SHMS->AddDetector(hod);
-
-  // Add trigger apparatus
-  THaApparatus* TRG = new THcTrigApp("T", "TRG");
-  gHaApps->Add(TRG);
-  // Add trigger detector to trigger apparatus
-  THcTrigDet* shms = new THcTrigDet("shms", "SHMS Trigger Information");
-  TRG->AddDetector(shms);
-
-  // Add handler for prestart event 125.
-  THcConfigEvtHandler* ev125 = new THcConfigEvtHandler("HC", "Config Event type 125");
-  gHaEvtHandlers->Add(ev125);
-
-  // Set up the analyzer - we use the standard one,
-  // but this could be an experiment-specific one as well.
-  // The Analyzer controls the reading of the data, executes
-  // tests/cuts, loops over Acpparatus's and PhysicsModules,
-  // and executes the output routines.
-  THcAnalyzer* analyzer = new THcAnalyzer;
-
-  // A simple event class to be output to the resulting tree.
-  // Creating your own descendant of THaEvent is one way of
-  // defining and controlling the output.
-  THaEvent* event = new THaEvent;
-
-  // Define the run(s) that we want to analyze.
-  // We just set up one, but this could be many.
-  THcRun* run = new THcRun( pathList, Form(RunFileNamePattern, RunNumber) );
-  // Set to read in Hall C run database parameters
-  run->SetRunParamClass("THcRunParameters");
-
-  // Eventually need to learn to skip over, or properly analyze
-  // the pedestal events
-  run->SetEventRange(1, MaxEvent);    // Physics Event number, does not
-                                      // include scaler or control events.
-  run->SetNscan(1);
-  run->SetDataRequired(0x7);
-  run->Print();
-
-  // Define the analysis parameters
-  TString ROOTFileName = Form(ROOTFileNamePattern, RunNumber);
-  analyzer->SetCountMode(2);    // 0 = counter is # of physics triggers
-                                // 1 = counter is # of all decode reads
-                                // 2 = counter is event number
-  analyzer->SetEvent(event);
-  analyzer->SetCrateMapFileName("MAPS/db_cratemap.dat");
-  analyzer->SetOutFile(ROOTFileName.Data());
-  analyzer->SetOdefFile("DEF-files/SHMS/TEST_STANDS/NGCER/pngcerana.def");
-  analyzer->SetCutFile("DEF-files/SHMS/TEST_STANDS/NGCER/pngcerana_cuts.def");    // optional
-
-  // File to record cuts accounting information
-  //analyzer->SetSummaryFile("summary_example.log");    // optional
-
-  // Start the actual analysis.
-  analyzer->Process(run);
-  // Create report file from template.
-  //analyzer->PrintReport(    // optional
-  //  "TEMPLATES/dcana.template",
-  //  Form("REPORT_OUTPUT/replay_hms_%05d.report", RunNumber)
-  //);
-}
